Bulk access to package and resource metadata
 - clean up and modularise code
 - we now depend on ckanext-scheming
 - output one CSV file per schema in the result set, using the order of
 fields in the schema
